added order_details through table, data to seed.js file, and seeded db

+++ server/db/models/Order.js
@@ -2,14 +2,12 @@ const Sequelize = require("sequelize");
 const db = require("../db");
 
 const Order = db.define("order", {
-  order_total: Sequelize.DECIMAL,
+  order_total: Sequelize.INTEGER,
   total_quantity: Sequelize.INTEGER,
-
   is_completed: {
     type: Sequelize.BOOLEAN,
     defaultValue: false,
   },
-
   date_placed: Sequelize.DATE,
   date_shipped: Sequelize.DATE,
 });
